完成结算系数表的增删改查

+++ scrapSettlement/DAL/Model/Coefficient.cs
@@ -17,7 +17,7 @@ namespace ScrapSettlement.DAL.Model
 
         public int CustomerID { get; set;  }
         public int ScrapID { get; set; }
-        public float Data { get; set; }
+        public Double Data { get; set; }
         public DateTime EffectDate { get; set; }
 
         public Nullable<DateTime> FailureDate { get; set; }
